newscard & commitcard moved to components
