+++ src/js/components/commitcard.js
@@ -1,4 +1,4 @@
-import { formatDateAsLongString } from './utils/datetime';
+import { formatDateAsLongString } from '../utils/datetime';
 
 export class CommitCard {
     constructor(api, template) {
